+++ packages/llm_chatgpt/lib/src/chatgpt_chat_repository.dart
@@ -196,4 +196,13 @@ class ChatGPTChatRepository extends LLMChatRepository {
         );
     }
   }
+
+  @override
+  Future<List<LLMEmbedding>> batchEmbed({
+    required String model,
+    required List<String> messages,
+    Map<String, dynamic> options = const {},
+  }) async {
+    return embed(model: model, messages: messages, options: options);
+  }
 }